Export SidelistMenu to new component
